feat: change selector to menu & update spacing/layout
