+++ app/client/src/PluginActionEditor/components/PluginActionResponse/components/QueryResponseTab/index.ts
@@ -1,0 +1,1 @@
+export { QueryResponseTab as default } from "./QueryResponseTab";
