+++ app/client/src/pages/Editor/QueryEditor/QueryDebuggerTabs.tsx
@@ -2,9 +2,7 @@ import type { BottomTab } from "components/editorComponents/EntityBottomTabs";
 import EntityBottomTabs from "components/editorComponents/EntityBottomTabs";
 import React, { useCallback, useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import styled from "styled-components";
 import { getErrorCount } from "selectors/debuggerSelectors";
-import { Text, TextType } from "@appsmith/ads-old";
 import { DEBUGGER_TAB_KEYS } from "components/editorComponents/Debugger/constants";
 import {
   DEBUGGER_ERRORS,
@@ -16,7 +14,6 @@ import DebuggerLogs from "components/editorComponents/Debugger/DebuggerLogs";
 import ErrorLogs from "components/editorComponents/Debugger/Errors";
 import Schema from "PluginActionEditor/components/PluginActionResponse/components/Schema";
 import type { ActionResponse } from "api/ActionAPI";
-import { isString } from "lodash";
 import type { SourceEntity } from "entities/AppsmithConsole";
 import type { Action } from "entities/Action";
 import QueryResponseTab from "PluginActionEditor/components/PluginActionResponse/components/QueryResponseTab";
@@ -36,17 +33,6 @@ import { actionResponseDisplayDataFormats } from "../utils";
 import { getIDEViewMode } from "selectors/ideSelectors";
 import { EditorViewMode } from "ee/entities/IDE/constants";
 import { IDEBottomView, ViewHideBehaviour } from "IDE";
-
-const ResultsCount = styled.div`
-  position: absolute;
-  right: ${(props) => props.theme.spaces[17] + 1}px;
-  top: 9px;
-  color: var(--ads-v2-color-fg);
-`;
-
-const ErrorText = styled(Text)`
-  color: var(--ads-v2-colors-action-error-label-default-fg);
-`;
 
 interface QueryDebuggerTabsProps {
   actionSource: SourceEntity;
@@ -71,9 +57,6 @@ function QueryDebuggerTabs({
   runErrorMessage,
   showSchema,
 }: QueryDebuggerTabsProps) {
-  // TODO: Fix this the next time the file is edited
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let output: Record<string, any>[] | null = null;
   const dispatch = useDispatch();
 
   const { open, responseTabHeight, selectedTab } = useSelector(
@@ -172,27 +155,6 @@ function QueryDebuggerTabs({
     }
   }, [currentActionConfig?.id]);
 
-  // Query is executed even once during the session, show the response data.
-  if (actionResponse) {
-    if (isString(actionResponse.body)) {
-      try {
-        // Try to parse response as JSON array to be displayed in the Response tab
-        output = JSON.parse(actionResponse.body);
-      } catch (e) {
-        // In case the string is not a JSON, wrap it in a response object
-        output = [
-          {
-            response: actionResponse.body,
-          },
-        ];
-      }
-    } else {
-      // TODO: Fix this the next time the file is edited
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      output = actionResponse.body as any;
-    }
-  }
-
   const setQueryResponsePaneHeight = useCallback(
     (height: number) => {
       dispatch(
@@ -276,21 +238,6 @@ function QueryDebuggerTabs({
       onHideClick={onToggle}
       setHeight={setQueryResponsePaneHeight}
     >
-      {output && !!output.length && (
-        <ResultsCount>
-          <Text data-testid="result-text" type={TextType.P3}>
-            Result:
-            {actionResponse?.isExecutionSuccess ? (
-              <Text type={TextType.H5}>{` ${output.length} Record${
-                output.length > 1 ? "s" : ""
-              }`}</Text>
-            ) : (
-              <ErrorText type={TextType.H5}>{" Error"}</ErrorText>
-            )}
-          </Text>
-        </ResultsCount>
-      )}
-
       <EntityBottomTabs
         isCollapsed={!open}
         onSelect={setSelectedResponseTab}
